// server/types/wind-conditions.ts
export type WindConditionRaw = {
  date: Date;
  windSpeed: number | null;
  windGusts: number | null;
  windDirection: number | null;
  waveHeight: number | null;
  wavePeriod: number | null;
  waveDirection: number | null;
  swellHeight: number | null;
  swellPeriod: number | null;
  swellDirection: number | null;
};